# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Church ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/account:
    delete:
      operationId: deleteOwnAccount
      tags: [auth]
      summary: Permanently delete the current account and personal data
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteOwnAccountRequest"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteOwnAccountResponse"
        "400":
          description: Invalid confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Last active administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List and filter user accounts (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, default: 20 } }
        - { in: query, name: status, schema: { $ref: "#/components/schemas/AccountStatus" } }
        - { in: query, name: role, schema: { $ref: "#/components/schemas/AccountRole" } }
        - { in: query, name: search, schema: { type: string } }
      responses:
        "200":
          description: Accounts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountsListResponse"
  /admin/accounts/member-options:
    get:
      operationId: listAccountMemberOptions
      tags: [accounts]
      summary: Search members and their account links (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
      responses:
        "200":
          description: Member choices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountMemberOptionsResponse"
  /admin/accounts/{id}/reject:
    post:
      operationId: rejectAccount
      tags: [accounts]
      summary: Reject a pending request without deleting it (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountReasonRequest"
      responses:
        "200":
          description: Rejected account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}/reopen:
    post:
      operationId: reopenAccount
      tags: [accounts]
      summary: Reopen a rejected request as pending (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Pending account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get an account (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}/approve:
    post:
      operationId: approveAccount
      tags: [accounts]
      summary: Approve a pending account (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveAccountRequest"
      responses:
        "200":
          description: Approved account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}/block:
    post:
      operationId: blockAccount
      tags: [accounts]
      summary: Temporarily block an account (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountReasonRequest"
      responses:
        "200":
          description: Blocked account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}/unblock:
    post:
      operationId: unblockAccount
      tags: [accounts]
      summary: Unblock an account (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Active account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}/revoke:
    post:
      operationId: revokeAccount
      tags: [accounts]
      summary: Revoke an account (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountReasonRequest"
      responses:
        "200":
          description: Revoked account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}/reactivate:
    post:
      operationId: reactivateAccount
      tags: [accounts]
      summary: Reactivate a revoked account (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Active account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}/role:
    patch:
      operationId: updateAccountRole
      tags: [accounts]
      summary: Change a member or leader role (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRoleRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/accounts/{id}/member-link:
    patch:
      operationId: updateAccountMemberLink
      tags: [accounts]
      summary: Link an account to a member (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountMemberLinkRequest"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    AccountStatus:
      type: string
      enum: [pending, rejected, active, blocked, revoked, deleting]
    AccountMemberOption:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, nullable: true }
        status: { type: string }
        linkedAccountId: { type: string, nullable: true }
        linkedAccountName: { type: string, nullable: true }
      required: [id, name, status]
    AccountMemberOptionsResponse:
      type: object
      properties:
        members:
          type: array
          items: { $ref: "#/components/schemas/AccountMemberOption" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [members, total, page, limit]
    AccountRole:
      type: string
      enum: [admin, leader, member]
    AdminAccount:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: "#/components/schemas/AccountRole" }
        status: { $ref: "#/components/schemas/AccountStatus" }
        emailVerifiedAt: { type: string, format: date-time, nullable: true }
        memberId: { type: string, nullable: true }
        memberLinkReviewedAt: { type: string, format: date-time, nullable: true }
        memberName: { type: string, nullable: true }
        statusReason: { type: string, nullable: true }
        requestedAt: { type: string, format: date-time }
        approvedAt: { type: string, format: date-time, nullable: true }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, email, name, role, status, requestedAt, createdAt, updatedAt]
    AccountStatusSummary:
      type: object
      properties:
        pending: { type: integer }
        rejected: { type: integer }
        active: { type: integer }
        blocked: { type: integer }
        revoked: { type: integer }
        deleting: { type: integer }
      required: [pending, rejected, active, blocked, revoked, deleting]
    AccountsListResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AdminAccount"
        summary:
          $ref: "#/components/schemas/AccountStatusSummary"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [accounts, summary, total, page, limit]
    AccountReasonRequest:
      type: object
      properties:
        reason: { type: string, minLength: 1 }
      required: [reason]
    ApproveAccountRequest:
      type: object
      properties:
        memberId: { type: string, nullable: true }
    UpdateAccountRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [member, leader]
      required: [role]
    UpdateAccountMemberLinkRequest:
      type: object
      properties:
        memberId: { type: string, nullable: true }
      required: [memberId]
    DeleteOwnAccountRequest:
      type: object
      properties:
        password: { type: string }
        confirmation: { type: string, enum: [EXCLUIR] }
      required: [password, confirmation]
    DeleteOwnAccountResponse:
      type: object
      properties:
        message: { type: string }
        deletionReference: { type: string }
      required: [message, deletionReference]
...