# lib/api-zod/src/generated/types/accountMemberOption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Church ERP API
 * OpenAPI spec version: 0.2.0
 */

export interface AccountMemberOption {
  id: string;
  name: string;
  email?: string | null;
  status: string;
  linkedAccountId?: string | null;
  linkedAccountName?: string | null;
}