# lib/api-zod/src/generated/types/accountMemberOptionsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Church ERP API
 * OpenAPI spec version: 0.2.0
 */
import type { AccountMemberOption } from "./accountMemberOption";

export interface AccountMemberOptionsResponse {
  members: AccountMemberOption[];
  total: number;
  page: number;
  limit: number;
}