# lib/api-zod/src/generated/types/accountStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Church ERP API
 * OpenAPI spec version: 0.2.0
 */

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus];

export const AccountStatus = {
  pending: "pending",
  rejected: "rejected",
  active: "active",
  blocked: "blocked",
  revoked: "revoked",
  deleting: "deleting",
} as const;