# lib/api-zod/src/generated/types/accountStatusSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Church ERP API
 * OpenAPI spec version: 0.2.0
 */

export interface AccountStatusSummary {
  pending: number;
  rejected: number;
  active: number;
  blocked: number;
  revoked: number;
  deleting: number;
}