# lib/api-zod/src/generated/types/adminAccount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Church ERP API
 * OpenAPI spec version: 0.2.0
 */
import type { AccountRole } from "./accountRole";
import type { AccountStatus } from "./accountStatus";

export interface AdminAccount {
  id: string;
  email: string;
  name: string;
  role: AccountRole;
  status: AccountStatus;
  emailVerifiedAt?: Date | null;
  memberId?: string | null;
  memberLinkReviewedAt?: Date | null;
  memberName?: string | null;
  statusReason?: string | null;
  requestedAt: Date;
  approvedAt?: Date | null;
  lastLoginAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}